refactor: modernize package with tsdx
