--- .babelrc
@@ -1,4 +1,0 @@
-{
-  "sourceMaps": "inline",
-  "presets": ["@babel/preset-env"]
-}
