+++ LICENSE.md
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) Christoph Fricke
+Copyright (c) 2019-present Christoph Fricke
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
