+++ package.json
@@ -10,36 +10,56 @@
     "input validation",
     "validation"
   ],
+  "repository": "github:christoph-fricke/use-input-validation",
   "main": "dist/index.js",
-  "typings": "typings/index.d.ts",
+  "module": "dist/use-input-validation.esm.js",
+  "typings": "dist/index.d.ts",
+  "sideEffects": false,
   "files": [
-    "dist/*",
-    "typings/*"
+    "dist",
+    "README.md",
+    "LICENSE.md"
   ],
   "scripts": {
-    "prepublishOnly": "babel ./src --out-dir ./dist",
-    "test": "jest --watch"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/christoph-fricke/use-input-validation.git"
-  },
-  "bugs": {
-    "url": "https://github.com/christoph-fricke/use-input-validation/issues"
+    "prepublishOnly": "tsdx build",
+    "build": "tsdx build",
+    "test": "tsdx test",
+    "test:watch": "tsdx test --watch",
+    "test:cov": "tsdx test --coverage",
+    "lint": "tsdx lint",
+    "format": "tsdx lint --fix",
+    "size": "size-limit",
+    "analyze": "size-limit --why"
   },
   "peerDependencies": {
     "react": ">=16.9.0"
   },
-  "devDependencies": {
-    "@babel/cli": "^7.5.5",
-    "@babel/core": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@testing-library/react-hooks": "^2.0.1",
-    "jest": "^24.9.0",
-    "react": "^16.9.0",
-    "react-test-renderer": "^16.9.0"
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
+    }
   },
-  "dependencies": {
-    "@types/react": ">=16.9.2"
+  "size-limit": [
+    {
+      "path": "dist/use-input-validation.cjs.production.min.js",
+      "limit": "1 KB"
+    },
+    {
+      "path": "dist/use-input-validation.esm.js",
+      "limit": "1 KB"
+    }
+  ],
+  "devDependencies": {
+    "@size-limit/preset-small-lib": "^4.6.0",
+    "@testing-library/react-hooks": "^3.4.2",
+    "@types/react": "^16.9.51",
+    "@types/react-dom": "^16.9.8",
+    "husky": "^4.3.0",
+    "react": "^16.13.1",
+    "react-test-renderer": "^16.13.1",
+    "size-limit": "^4.6.0",
+    "tsdx": "^0.14.0",
+    "tslib": "^2.0.2",
+    "typescript": "^4.0.3"
   }
 }
